+++ database/migrations/2020_03_26_234133_create_users_roles_table.php
@@ -13,17 +13,17 @@ class CreateUsersRolesTable extends Migration
      */
     public function up()
     {
-        Schema::create('users_roles', function (Blueprint $table) {
-            $table->id();
-            $table->integer('user_id')->unsigned()->index();
-            $table->integer('role_id')->unsigned()->index();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
-            $table->timestamps();
-
-        });
+        if (!Schema::hasTable('users_roles')) {
+            Schema::create('users_roles', function (Blueprint $table) {
+                $table->id();
+                $table->integer('user_id')->unsigned()->index();
+                $table->integer('role_id')->unsigned()->index();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
-
     /**
      * Reverse the migrations.
      *
